Use WriteConnectionSecretToRef instead of custom secretRef
This should avoid some confusion and duplication of a secret reference.
