+++ docs/modules/ROOT/examples/cloudscale_objectsuser.yaml
@@ -4,10 +4,10 @@ metadata:
   name: my-cloudscale-user
 spec:
   forProvider:
-    secretRef:
-      name: my-cloudscale-user-credentials
-      namespace: default
     tags:
       key: value
   providerConfigRef:
     name: provider-config
+  writeConnectionSecretToRef:
+    name: my-cloudscale-user-credentials
+    namespace: default
